+++ Wflow/src/routing/initialize_routing.jl
@@ -46,13 +46,13 @@ function initialize_subsurface_flow(
     n_cells = length(indices)
 
     lens = lens_input_parameter(config, "land_surface__elevation"; optional = false)
-    elevation = ncread(dataset, config, lens; sel = indices, type = Float64)
+    elevation = ncread(dataset, config, lens; sel = indices, type = Float)
 
     # unconfined aquifer
     if do_constanthead
         constant_head = ConstantHead(dataset, config, indices)
     else
-        variables = ConstantHeadVariables(; head = Float64[])
+        variables = ConstantHeadVariables(; head = Float[])
         constant_head = ConstantHead(; variables, index = Int64[])
     end
 
@@ -66,7 +66,7 @@ function initialize_subsurface_flow(
     end
 
     bottom = elevation .- soil.parameters.soilthickness ./ 1000.0
-    conductance = zeros(Float64, connectivity.nconnection)
+    conductance = zeros(Float, connectivity.nconnection)
     aquifer = UnconfinedAquifer(
         dataset,
         config,
